[FIX] Performances issues with colisions and indexed zipper

+++ shared/Systems.cpp
@@ -45,9 +45,12 @@ void update_grace(Registry &r,
 sparse_array<SpawnGrace> &graces)
 {
     auto time = GetTimePoint();
+    auto graces_size = graces.size();
 
-    for (auto &&[ind, grace]: indexed_zipper(graces)) {
-        if (grace->creation_time + grace->time >= time) {
+	for (size_t ind = 0; ind != graces_size; ++ind) {
+        if (!graces[ind])
+            continue;
+        if (graces[ind]->creation_time + graces[ind]->time >= time) {
             r.remove_component<SpawnGrace>(ind);
         }
     }
@@ -55,36 +58,42 @@ sparse_array<SpawnGrace> &graces)
 
 void colision(Registry &r,
 sparse_array<Position> &positions, 
-sparse_array<Size> &size, 
+sparse_array<Size> &sizes,
 sparse_array<SpawnGrace> &grace, 
-sparse_array<Damages> &dam, 
-sparse_array<Health> &helth)
+sparse_array<Damages> &dmgs,
+sparse_array<Health> &healths)
 {
-    for (auto &&[ind, pos, siz, dama, halth]: indexed_zipper(positions, size, dam, helth)) {
+    auto pos_size = positions.size();
+	for (size_t ind = 0; ind != pos_size; ++ind) {
+        if (!(positions[ind] && sizes[ind] && dmgs[ind] && healths[ind]))
+            continue;
+
         if (grace[ind].has_value()) {
             continue;
         }
-        for (auto &&[ind2, pos2, siz2, dama2, halth2]: indexed_zipper(positions, size, dam, helth)) {
-            if (!positions[ind]) { // need to recheck because damages may have kill the entity
-                continue;
-            }
+		for (size_t ind2 = 0; ind2 != pos_size; ++ind2) {
+			if (!(positions[ind2] && sizes[ind2] && dmgs[ind2] && healths[ind2]))
+				continue;
             if (ind2 <= ind || grace[ind2].has_value()) {
                 continue;
             }
+            if (!positions[ind]) { // need to recheck because damages may have kill the entity
+                continue;
+            }
             if (positions[ind]->pos_X >
-				positions[ind2]->pos_X + size[ind2]->size_X)
+				positions[ind2]->pos_X + sizes[ind2]->size_X)
 				continue;
             else if (
-                positions[ind]->pos_Y > positions[ind2]->pos_Y + size[ind2]->size_Y)
+                positions[ind]->pos_Y > positions[ind2]->pos_Y + sizes[ind2]->size_Y)
                 continue;
             else if (
-                positions[ind2]->pos_X > positions[ind]->pos_X + size[ind]->size_X)
+                positions[ind2]->pos_X > positions[ind]->pos_X + sizes[ind]->size_X)
                 continue;
             else if (
-                positions[ind2]->pos_Y > positions[ind]->pos_Y + size[ind]->size_Y)
+                positions[ind2]->pos_Y > positions[ind]->pos_Y + sizes[ind]->size_Y)
                 continue;
             else
-                damages(r, helth, dam, ind, ind2);
+                damages(r, healths, dmgs, ind, ind2);
         }
     }
 }
